+++ scripts/load_vocabularies.py
@@ -1,6 +1,6 @@
 import os
 from os import path
-
+import argparse
 import requests
 
 from db_utils import base, get_auth
@@ -12,24 +12,24 @@ ROOT = './dump'
 VOCAB = path.join(ROOT, 'vocabularies')
 INTERNAL_ROOT = '/opt/graphdb/home/graphdb-import'
 
-
 C_TYPE = {
     'rdf': 'application/rdf+xml',
     'ttl': 'text/turtle'
 }
 
 
-def upload_in(file_path, content_type, graph_name):
+def upload_in(file_path, content_type, graph_name, keep_data=False):
     # clear graph
     headers = {'Authorization': get_auth()}
     params = (('graph', graph_name),)
 
-    response = requests.delete(f'{base}/repositories/odeuropa/rdf-graphs/service',
-                               headers=headers, params=params)
-    if response.status_code != 204:
-        print(response.status_code)
-        print(response.content)
-        return
+    if not keep_data:
+        response = requests.delete(f'{base}/repositories/odeuropa/rdf-graphs/service',
+                                   headers=headers, params=params)
+        if response.status_code != 204:
+            print(response.status_code)
+            print(response.content)
+            return
 
     # upload new resource
     headers['Content-Type'] = content_type
@@ -44,13 +44,20 @@ def upload_in(file_path, content_type, graph_name):
         print(response.content)
 
 
-for filename in sorted(os.listdir(VOCAB)):
-    name, ext = filename.rsplit('.')
-    if ext == 'DS_Store':
-        continue
-    print('- ' + name)
-    if C_TYPE[ext]:
-        upload_in(path.join(VOCAB, filename), C_TYPE[ext], path.join(VOCAB_GRAPH, name))
-    else:
-        continue
-print('completed')
+def load_vocabularies(keep_data=False):
+    for filename in sorted(os.listdir(VOCAB)):
+        name, ext = filename.rsplit('.')
+        if ext == 'DS_Store':
+            continue
+        print('- ' + name)
+        if C_TYPE[ext]:
+            upload_in(path.join(VOCAB, filename), C_TYPE[ext], path.join(VOCAB_GRAPH, name), keep_data)
+        else:
+            continue
+    print('completed')
+
+
+parser = argparse.ArgumentParser(description='Load dump in a graph.')
+parser.add_argument('-k', '--keep_data', description='Skip the deletion of the graph', action='store_true')
+args = parser.parse_args()
+load_vocabularies(args.keep_data)
